% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor.R
\name{pcor}
\alias{pcor}
\title{Projection Correlation}
\usage{
pcor(X, Y)
}
\arguments{
\item{X}{A numeric matrix, n*p, each row of the matrix is i.i.d generated from one vector}

\item{Y}{A numeric matrix, n*q, each row of the matrix is i.i.d generated from the other vector}
}
\value{
The projection correlation of X and Y
}
\description{
Calculate the projection correlation of two random vectors. Two vectors can be with different dimensions, but with equal sample sizes.
}
\examples{
X = matrix(rnorm(100*34,1),100,34)
Y = matrix(rnorm(100*62,1),100,62)
pcor(X,Y)
}
\references{
L.Zhu, K.Xu, R.Li, W.Zhong(2017). Projection correlation between two random vectors. Biometrika, Volume 104, Pages 829–843. https://doi.org/10.1093/biomet/asx043
}
\seealso{
\code{\link{pcov}} \code{\link{pcor.test}}
}
