% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor.R
\name{pcor.test}
\alias{pcor.test}
\title{Projection Correlatoin Permutation Test}
\usage{
pcor.test(X, Y)
}
\arguments{
\item{X}{A numeric matrix, n*p, each row of the matrix is i.i.d generated from one vector}

\item{Y}{A numeric matrix, n*q, each row of the matrix is i.i.d generated from the other vector}
}
\value{
\code{pcor.value} projection correlation of X and Y

\code{p.value} the p-value under the null hypothesis two vectors are independent
}
\description{
Return the test result of projection correlation test. Test whether two vectors are independnet. Two vectors can be with different dimensions, but with equal sample sizes.
}
\examples{
X = matrix(rnorm(10*3,1),10,3)
Y = matrix(rnorm(10*6,1),10,6)
pcor.test(X,Y)
}
\references{
L.Zhu, K.Xu, R.Li, W.Zhong(2017). Projection correlation between two random vectors. Biometrika, Volume 104, Pages 829–843. https://doi.org/10.1093/biomet/asx043
}
\seealso{
\code{\link{pcov}} \code{\link{pcor}}
}
