% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor.R
\name{pcov}
\alias{pcov}
\title{Projection Covariance}
\usage{
pcov(X, Y, estimation.method = "u", parallel = FALSE)
}
\arguments{
\item{X}{A numeric matrix, n*p, each row of the matrix is i.i.d generated from one vector}

\item{Y}{A numeric matrix, n*q, each row of the matrix is i.i.d generated from the other vector}

\item{estimation.method}{A character, "u" or "v". When "u", it will use U statistics to estimate projection covariance and other statistics. When "v", it will use V statistics instead.}

\item{parallel}{A logical, TRUE or FALSE. When TRUE, it will compute parallel, which is faster when sample is larger than around 100. Default is FALSE.}
}
\value{
The projection covariance of X and Y
}
\description{
Calculate the projection covariance of two random vectors. Two vectors can be with different dimensions, but with equal sample sizes.
}
\examples{
X = matrix(rnorm(100*34,1),100,34)
Y = matrix(rnorm(100*62,1),100,62)
pcov(X, Y, estimation.method="u", parallel=FALSE)
}
\references{
L.Zhu, K.Xu, R.Li, W.Zhong(2017). Projection correlation between two random vectors. Biometrika, Volume 104, Pages 829-843. https://doi.org/10.1093/biomet/asx043
}
\seealso{
\code{\link{pcor}} \code{\link{pcor.test}}
}
